% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_file_path_input.R
\name{tar_input_file_path}
\alias{tar_input_file_path}
\title{Workflow input using file paths}
\usage{
tar_input_file_path(name, mzML_files, sample_information)
}
\arguments{
\item{name}{Symbol. The name for the collection of targets. This serves as a prefix for target names.}

\item{mzML_files}{A characer vector of mzML file paths.}

\item{sample_information}{A tibble containing the sample information. See details for the specifications.}
}
\value{
A list of target objects specifically for analysis pipeline input using file paths.
}
\description{
Targets for workflow inputs using file paths
}
\details{
The tibble containing sample information should at least contain the following columns:
\itemize{
\item \code{fileOrder} - the numeric file order of the input files when order alphabetically as returned by list.files().
\item \code{injOrder} - the sample injection order during MS analysis
\item \code{fileName} - the sample file name
\item \code{batch} - the analytical batch number
\item \code{block} - the randomised block number
\item \code{name} - the sample name
\item \code{class} - the sample class name
}

The file names contained within the argument \code{mzML_files} should match in name and order to those within the \code{fileName} sample information column.
}
\examples{
## File path input 
targets::tar_dir({
targets::tar_script({
    library(hrmtargets)
    file_paths <- metaboData::filePaths('FIE-HRMS','UrineTechnical',ask = FALSE)
    sample_info <- metaboData::runinfo('FIE-HRMS','UrineTechnical',ask = FALSE)
    
    list(
        tar_input_file_path(test,
                            file_paths,
                            sample_info)
    )
})
targets::tar_make()
})
}
