% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_spectral_processing.R
\name{tar_spectral_processing}
\alias{tar_spectral_processing}
\title{Spectral processing}
\usage{
tar_spectral_processing(
  name,
  type = c("FIE-HRMS", "RP-LC-HRMS", "NP-LC-HRMS"),
  ...
)
}
\arguments{
\item{name}{Symbol. The name for the collection of targets. This serves as a prefix for target names.}

\item{type}{The type of analytical technique.}

\item{...}{Arguments to pass to the relevant target factory. See details.}
}
\value{
A list of target objects for spectral processing for the specified technique.
}
\description{
Targets for spectral processing.
}
\details{
Technique \code{FIE-HRMS} will return targets using \code{tar_spectral_binning()}.
Techniques \code{RP-LC-HRMS} and \code{NP-LC-HRMS} will return targets using \code{tar_profiling()}.
}
\examples{
\dontrun{
## Perform spectral processing using technique `FIE-HRMS` 
targets::tar_dir({
    targets::tar_script({
        library(hrmtargets)
        name <- rlang::expr(example)
        
        list(
            tar_input_piggyback(!!name,
                                'FIE-HRMS_BdistachyonTechnical',
                                repo = 'jasenfinch/metaboData'),
            tar_spectral_processing(!!name,
                                    type = 'FIE-HRMS')
        )
    })
    targets::tar_make()
    targets::tar_read(example_results_spectral_processing)
    targets::tar_read(example_plot_fingerprint)
})
}
}
